[sw/lib] Move print.h and log.h into lib/runtime
