+++ sw/device/boot_rom/meson.build
@@ -43,7 +43,7 @@ foreach device_name, device_lib : sw_lib_arch_core_devices
     dependencies: [
       chip_info_h,
       sw_lib_runtime_hart,
-      sw_lib_base_print,
+      sw_lib_runtime_print,
       sw_lib_flash_ctrl,
       sw_lib_pinmux,
       sw_lib_dif_gpio,
@@ -51,7 +51,7 @@ foreach device_name, device_lib : sw_lib_arch_core_devices
       sw_lib_hmac,
       sw_lib_mmio,
       sw_lib_uart,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       top_earlgrey,
       device_lib,
       sw_lib_testing_test_status,
