+++ sw/device/examples/demos.c
@@ -7,11 +7,11 @@
 #include <stdbool.h>
 
 #include "sw/device/lib/arch/device.h"
-#include "sw/device/lib/base/log.h"
 #include "sw/device/lib/dif/dif_gpio.h"
 #include "sw/device/lib/dif/dif_spi_device.h"
 #include "sw/device/lib/runtime/check.h"
 #include "sw/device/lib/runtime/hart.h"
+#include "sw/device/lib/runtime/log.h"
 #include "sw/device/lib/uart.h"
 
 void demo_gpio_startup(dif_gpio_t *gpio) {
