+++ sw/device/examples/hello_usbdev/meson.build
@@ -18,7 +18,7 @@ foreach device_name, device_lib : sw_lib_arch_core_devices
       sw_lib_irq,
       sw_lib_dif_spi_device,
       sw_lib_uart,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_usb,
       riscv_crt,
       sw_lib_irq_handlers,
