+++ sw/device/examples/meson.build
@@ -11,7 +11,7 @@ sw_examples_demos = declare_dependency(
       sw_lib_dif_gpio,
       sw_lib_dif_spi_device,
       sw_lib_uart,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
     ],
   )
 )
