+++ sw/device/lib/base/meson.build
@@ -30,31 +30,3 @@ sw_lib_mmio = declare_dependency(
   )
 )
 
-# Basic printing library (sw_lib_base_print)
-sw_lib_base_print = declare_dependency(
-  link_with: static_library(
-    'base_print_ot',
-    sources: ['print.c'],
-  )
-)
-
-# Generic logging library (sw_lib_base_log)
-sw_lib_base_log = declare_dependency(
-  link_with: static_library(
-    'base_log_ot',
-    sources: ['log.c'],
-    dependencies: [
-      sw_lib_base_print,
-    ],
-  )
-)
-
-test('sw_lib_base_print_test', executable(
-  'sw_lib_base_print_test',
-  sources: ['print.c', 'print_test.cc'],
-  dependencies: [
-    sw_vendor_gtest,
-  ],
-  native: true,
-))
-
