+++ sw/device/lib/runtime/log.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef OPENTITAN_SW_DEVICE_LIB_BASE_LOG_H_
-#define OPENTITAN_SW_DEVICE_LIB_BASE_LOG_H_
+#ifndef OPENTITAN_SW_DEVICE_LIB_RUNTIME_LOG_H_
+#define OPENTITAN_SW_DEVICE_LIB_RUNTIME_LOG_H_
 
 #include <stdbool.h>
 #include <stdint.h>
@@ -182,4 +182,4 @@ void base_log_internal_dv(const log_fields_t *log, uint32_t nargs, ...);
  */
 #define LOG_FATAL(...) LOG(kLogSeverityFatal, __VA_ARGS__)
 
-#endif  // OPENTITAN_SW_DEVICE_LIB_BASE_LOG_H_
+#endif  // OPENTITAN_SW_DEVICE_LIB_RUNTIME_LOG_H_
