+++ sw/device/lib/testing/meson.build
@@ -9,7 +9,7 @@ sw_lib_testing_test_status = declare_dependency(
     sources: ['test_status.c'],
     dependencies: [
       sw_lib_mmio,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_runtime_hart,
     ],
   )
@@ -37,7 +37,7 @@ if coverage
       dependencies: [
         sw_lib_mem,
         sw_lib_uart,
-        sw_lib_base_log,
+        sw_lib_runtime_log,
       ],
     ),
   )
@@ -55,7 +55,7 @@ sw_lib_testing_test_main = declare_dependency(
     sources: ['test_main.c'],
     dependencies: [
       sw_lib_uart,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_testing_test_status,
       sw_lib_testing_test_coverage,
     ],
