+++ sw/device/riscv_compliance_support/meson.build
@@ -13,7 +13,7 @@ foreach device_name, device_lib : sw_lib_arch_core_devices
     sources: ['support.c'],
     dependencies: [
       sw_lib_uart,
-      sw_lib_base_print,
+      sw_lib_runtime_print,
       sw_lib_irq_handlers,
       sw_lib_mem,
       device_lib,
