+++ sw/device/tests/dif/meson.build
@@ -123,7 +123,7 @@ dif_plic_sanitytest_lib = declare_dependency(
       sw_lib_dif_plic,
       sw_lib_irq,
       sw_lib_mmio,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_runtime_hart,
       sw_lib_testing_test_status,
     ],
@@ -150,7 +150,7 @@ dif_rv_timer_sanitytest_lib = declare_dependency(
     sources: ['dif_rv_timer_sanitytest.c'],
     dependencies: [
       sw_lib_dif_rv_timer,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_mmio,
       sw_lib_runtime_hart,
       sw_lib_irq,
@@ -165,7 +165,7 @@ dif_hmac_sanitytest_lib = declare_dependency(
     sources: ['dif_hmac_sanitytest.c'],
     dependencies: [
       sw_lib_dif_hmac,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_mmio,
       sw_lib_runtime_hart,
     ],
