+++ sw/device/tests/sim_dv/gpio_test.c
@@ -5,7 +5,6 @@
 #include "sw/device/lib/dif/dif_gpio.h"
 
 #include "sw/device/lib/arch/device.h"
-#include "sw/device/lib/base/log.h"
 #include "sw/device/lib/base/mmio.h"
 #include "sw/device/lib/dif/dif_plic.h"
 #include "sw/device/lib/handler.h"
@@ -13,6 +12,7 @@
 #include "sw/device/lib/pinmux.h"
 #include "sw/device/lib/runtime/check.h"
 #include "sw/device/lib/runtime/hart.h"
+#include "sw/device/lib/runtime/log.h"
 #include "sw/device/lib/testing/test_main.h"
 #include "sw/device/lib/testing/test_status.h"
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
