+++ sw/device/tests/sim_dv/meson.build
@@ -12,7 +12,7 @@ uart_tx_rx_test_lib = declare_dependency(
       sw_lib_irq,
       sw_lib_mmio,
       sw_lib_uart,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_runtime_hart,
       sw_lib_testing_test_status,
       top_earlgrey,
@@ -30,7 +30,7 @@ gpio_test_lib = declare_dependency(
       sw_lib_irq,
       sw_lib_mmio,
       sw_lib_pinmux,
-      sw_lib_base_log,
+      sw_lib_runtime_log,
       sw_lib_runtime_hart,
       sw_lib_testing_test_status,
       top_earlgrey,
